Rename Main.java to FizzBuzzJazz.java
Java program that makes a method called PrintNumbers and run it from main  The method prints the numbers from 1 to 100. For multiples of 3, print "Fizz" instead of the number, for multiples of 5, print "Buzz". For numbers which are multiples of both 3 and 5, print "FizzBuzz". Additionally, for multiples of 7, print "Jazz".